test: move username URL tests to their own module
test_urls_without_admin.py is about the admin toggle in INSTALLED_APPS.
The hyphenated-username and format-suffix classes test the username URL
pattern instead, so they now live in test_username_urls.py. Pure move,
no test changes.
